refactor: change changelog class and add validate
- Move `JiraChangelog` and `JiraChangeItem` to `models/jira/common.py`
- Rename `JiraChangeItem` to `JiraChangelogItem`
- Change `JiraChangelog.id` from `str` to `int`
- Add a validation that `JiraChangelog.items` should be always sorted
  in ascending in id, which is convenient for user
  - It is alreadly sorted in api calls (ascending or decreasing) so
    no performance loss
